requirements.txt
# Requirements for Virtual Mouse Lab
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
biopython>=1.81
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0